Fix doc and rename MailTo -> HandleMailToClick
