--- network/src/main/java/MailTo.java
@@ -1,87 +1,0 @@
-/*
- *  Copyright 2020, TeamDev. All rights reserved.
- *
- *  Redistribution and use in source and/or binary forms, with or without
- *  modification, must retain the above copyright notice and the following
- *  disclaimer.
- *
- *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- *  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- *  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- *  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- *  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- *  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- *  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-import com.teamdev.jxbrowser.browser.Browser;
-import com.teamdev.jxbrowser.engine.Engine;
-import com.teamdev.jxbrowser.engine.EngineOptions;
-import com.teamdev.jxbrowser.net.UrlRequest;
-import com.teamdev.jxbrowser.net.callback.BeforeUrlRequestCallback;
-import com.teamdev.jxbrowser.net.callback.BeforeUrlRequestCallback.Response;
-import com.teamdev.jxbrowser.view.swing.BrowserView;
-
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.WindowAdapter;
-import java.awt.event.WindowEvent;
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
-
-import static com.teamdev.jxbrowser.engine.RenderingMode.HARDWARE_ACCELERATED;
-import static java.awt.Desktop.isDesktopSupported;
-
-/**
- * This example demonstrates how to open a default desktop mail client by clicking {@code mailto:} links.
- */
-public final class MailTo {
-
-    private static final String HTML = "<a href=\"mailto:test@example.com\">Click to mail</a>";
-
-    public static void main(String[] args) {
-        Engine engine = Engine.newInstance(
-                EngineOptions.newBuilder(HARDWARE_ACCELERATED).build());
-        Browser browser = engine.newBrowser();
-
-        engine.network().set(BeforeUrlRequestCallback.class, params -> {
-            Desktop desktop = Desktop.getDesktop();
-            if (isDesktopSupported() && desktop.isSupported(Desktop.Action.MAIL)) {
-                try {
-                    UrlRequest urlRequest = params.urlRequest();
-                    String url = urlRequest.url();
-                    if (url.contains("mailto:")) {
-                        desktop.mail(new URI(url));
-                    }
-                } catch (URISyntaxException | IOException e) {
-                    throw new RuntimeException(e);
-                }
-            }
-            return Response.cancel();
-        });
-
-        SwingUtilities.invokeLater(() -> {
-            BrowserView view = BrowserView.newInstance(browser);
-
-            JFrame frame = new JFrame("Mail");
-            frame.addWindowListener(new WindowAdapter() {
-                @Override
-                public void windowClosing(WindowEvent e) {
-                    engine.close();
-                }
-            });
-            frame.setDefaultCloseOperation(WindowConstants.DISPOSE_ON_CLOSE);
-            frame.add(view, BorderLayout.CENTER);
-            frame.setSize(800, 600);
-            frame.setLocationRelativeTo(null);
-            frame.setVisible(true);
-
-            browser.mainFrame().ifPresent(mainFrame -> mainFrame.loadHtml(HTML));
-        });
-    }
-}
